Stop using an external fileid program
The symbolic-debuginfo crate has everything required to replace it
already.

At the same time, look for breakpad symbols in the directory
they're supposed to be based on their fileid, rather than assuming that
if there's only one subdirectory, it's the right one.
This means breakpad syms directories for tests need to be adjusted,
which they are by taking the id as appearing on the MODULE line in the
sym file. This validates that obj.debug_id().breakpad() returns the
expected value.

Also change the instructions to generate the breakpad symbols for
examples, using https://github.com/mozilla/dump_syms, which simplifies
things, albeit modifies the linux example file.